+++ gax/CHANGELOG.md
@@ -4,6 +4,19 @@
 
 [1]: https://www.npmjs.com/package/gax-nodejs?activeTab=versions
 
+## [4.5.0](https://github.com/googleapis/gax-nodejs/compare/google-gax-v4.4.1...google-gax-v4.5.0) (2024-10-25)
+
+
+### Features
+
+* adds unit tests for sync and async reads of a pipeline transformation  ([#1663](https://github.com/googleapis/gax-nodejs/issues/1663)) ([c7995c4](https://github.com/googleapis/gax-nodejs/commit/c7995c40cd6552daa51a94e3b740b42e1da69285))
+
+
+### Bug Fixes
+
+* add dependency to fix tests ([#1658](https://github.com/googleapis/gax-nodejs/issues/1658)) ([564e9e7](https://github.com/googleapis/gax-nodejs/commit/564e9e75bfa85f45cef58fd178d27474752d0c65))
+* overhaul of server streaming retries ([#1653](https://github.com/googleapis/gax-nodejs/issues/1653)) ([629f4c4](https://github.com/googleapis/gax-nodejs/commit/629f4c409d6df8130662d5d27ada7fd743a3492a))
+
 ## [4.4.1](https://github.com/googleapis/gax-nodejs/compare/google-gax-v4.4.0...google-gax-v4.4.1) (2024-09-04)